[NFC] Rename TransformDialectExtensions to TransformExtensions
