+++ compiler/src/iree/compiler/Codegen/Common/TransformExtensions/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/Codegen/Common/TransformDialectExtensions/BUILD   #
+# compiler/src/iree/compiler/Codegen/Common/TransformExtensions/BUILD          #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
