+++ compiler/src/iree/compiler/Codegen/Interfaces/BUILD
@@ -45,10 +45,10 @@ iree_compiler_cc_library(
         ":BufferizationInterfaces",
         ":PartitionableLoopsInterface",
         ":ProcessorOpInterfaces",
-        "//compiler/src/iree/compiler/Codegen/Common/TransformDialectExtensions:TransformDialectCommonExtensions",
-        "//compiler/src/iree/compiler/Codegen/LLVMCPU/TransformDialectExtensions:TransformDialectLLVMCPUExtensions",
-        "//compiler/src/iree/compiler/Codegen/LLVMGPU/TransformDialectExtensions:TransformDialectLLVMGPUExtensions",
-        "//compiler/src/iree/compiler/Dialect/Flow/TransformDialectExtensions:TransformDialectFlowExtensions",
+        "//compiler/src/iree/compiler/Codegen/Common/TransformExtensions:TransformDialectCommonExtensions",
+        "//compiler/src/iree/compiler/Codegen/LLVMCPU/TransformExtensions:TransformDialectLLVMCPUExtensions",
+        "//compiler/src/iree/compiler/Codegen/LLVMGPU/TransformExtensions:TransformDialectLLVMGPUExtensions",
+        "//compiler/src/iree/compiler/Dialect/Flow/TransformExtensions:TransformDialectFlowExtensions",
         "//llvm-external-projects/iree-dialects:IREELinalgExtTransformOps",
         "//llvm-external-projects/iree-dialects:IREELinalgTransformDialect",
         "@llvm-project//mlir:LinalgTransformOps",
