+++ compiler/src/iree/compiler/Codegen/LLVMGPU/TransformExtensions/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/Dialect/Flow/TransformDialectExtensions/BUILD     #
+# compiler/src/iree/compiler/Codegen/LLVMGPU/TransformExtensions/BUILD         #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -12,34 +12,40 @@ iree_add_all_subdirs()
 
 iree_tablegen_library(
   NAME
-    TransformDialectFlowExtensionsOpGen
+    TransformDialectLLVMGPUExtensionsOpGen
   TD_FILE
-    "TransformDialectFlowExtensionsOps.td"
+    "TransformDialectLLVMGPUExtensionsOps.td"
   OUTS
-    --gen-op-decls TransformDialectFlowExtensionsOps.h.inc
-    --gen-op-defs TransformDialectFlowExtensionsOps.cpp.inc
+    --gen-op-decls TransformDialectLLVMGPUExtensionsOps.h.inc
+    --gen-op-defs TransformDialectLLVMGPUExtensionsOps.cpp.inc
 )
 
 iree_cc_library(
   NAME
-    TransformDialectFlowExtensions
+    TransformDialectLLVMGPUExtensions
   HDRS
-    "TransformDialectFlowExtensions.h"
-    "TransformDialectFlowExtensionsOps.h.inc"
+    "TransformDialectLLVMGPUExtensions.h"
+    "TransformDialectLLVMGPUExtensionsOps.h.inc"
   SRCS
-    "TransformDialectFlowExtensions.cpp"
-    "TransformDialectFlowExtensionsOps.cpp.inc"
+    "TransformDialectLLVMGPUExtensions.cpp"
+    "TransformDialectLLVMGPUExtensionsOps.cpp.inc"
   DEPS
-    ::TransformDialectFlowExtensionsOpGen
+    ::TransformDialectLLVMGPUExtensionsOpGen
     IREEDialectsTransforms
     IREELinalgTransformDialect
     MLIRArithmeticDialect
+    MLIRFuncDialect
+    MLIRGPUOps
     MLIRIR
+    MLIRMemRefDialect
+    MLIRPDLDialect
+    MLIRPass
     MLIRSCFDialect
-    MLIRTensorDialect
     MLIRTransformDialect
-    MLIRTransformUtils
-    iree::compiler::Dialect::Flow::IR
+    MLIRTransforms
+    MLIRVectorDialect
+    MLIRVectorTransforms
+    iree::compiler::Dialect::HAL::IR
   PUBLIC
 )
