+++ compiler/src/iree/compiler/Dialect/Flow/TransformExtensions/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/Codegen/LLVMGPU/TransformDialectExtensions/BUILD  #
+# compiler/src/iree/compiler/Dialect/Flow/TransformExtensions/BUILD            #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -12,40 +12,34 @@ iree_add_all_subdirs()
 
 iree_tablegen_library(
   NAME
-    TransformDialectLLVMGPUExtensionsOpGen
+    TransformDialectFlowExtensionsOpGen
   TD_FILE
-    "TransformDialectLLVMGPUExtensionsOps.td"
+    "TransformDialectFlowExtensionsOps.td"
   OUTS
-    --gen-op-decls TransformDialectLLVMGPUExtensionsOps.h.inc
-    --gen-op-defs TransformDialectLLVMGPUExtensionsOps.cpp.inc
+    --gen-op-decls TransformDialectFlowExtensionsOps.h.inc
+    --gen-op-defs TransformDialectFlowExtensionsOps.cpp.inc
 )
 
 iree_cc_library(
   NAME
-    TransformDialectLLVMGPUExtensions
+    TransformDialectFlowExtensions
   HDRS
-    "TransformDialectLLVMGPUExtensions.h"
-    "TransformDialectLLVMGPUExtensionsOps.h.inc"
+    "TransformDialectFlowExtensions.h"
+    "TransformDialectFlowExtensionsOps.h.inc"
   SRCS
-    "TransformDialectLLVMGPUExtensions.cpp"
-    "TransformDialectLLVMGPUExtensionsOps.cpp.inc"
+    "TransformDialectFlowExtensions.cpp"
+    "TransformDialectFlowExtensionsOps.cpp.inc"
   DEPS
-    ::TransformDialectLLVMGPUExtensionsOpGen
+    ::TransformDialectFlowExtensionsOpGen
     IREEDialectsTransforms
     IREELinalgTransformDialect
     MLIRArithmeticDialect
-    MLIRFuncDialect
-    MLIRGPUOps
     MLIRIR
-    MLIRMemRefDialect
-    MLIRPDLDialect
-    MLIRPass
     MLIRSCFDialect
+    MLIRTensorDialect
     MLIRTransformDialect
-    MLIRTransforms
-    MLIRVectorDialect
-    MLIRVectorTransforms
-    iree::compiler::Dialect::HAL::IR
+    MLIRTransformUtils
+    iree::compiler::Dialect::Flow::IR
   PUBLIC
 )
