+++ compiler/src/iree/compiler/Dialect/Flow/TransformExtensions/TransformDialectFlowExtensions.h
@@ -10,7 +10,7 @@
 #include "mlir/Dialect/Transform/IR/TransformDialect.h"
 
 #define GET_OP_CLASSES
-#include "iree/compiler/Dialect/Flow/TransformDialectExtensions/TransformDialectFlowExtensionsOps.h.inc"
+#include "iree/compiler/Dialect/Flow/TransformExtensions/TransformDialectFlowExtensionsOps.h.inc"
 
 namespace mlir {
 class DialectRegistry;
