+++ llvm-external-projects/iree-dialects/include/iree-dialects-c/Dialects.h
@@ -43,7 +43,7 @@ MLIR_CAPI_EXPORTED void mlirIREELinalgTransformRegisterPasses();
 MLIR_DECLARE_CAPI_DIALECT_REGISTRATION(Transform, transform);
 
 MLIR_CAPI_EXPORTED void
-ireeRegisterTransformDialectExtensions(MlirContext context);
+ireeRegisterTransformExtensions(MlirContext context);
 
 /// Register all passes for the transform dialect.
 MLIR_CAPI_EXPORTED void mlirIREETransformRegisterPasses();
